+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Next (TBD)
 
 * fix issue where the sequence number of a sentinel scene id can be two digit (ref: https://github.com/cogeotiff/rio-tiler-pds/pull/35)
+* fix issue where `utm` is only one sigle digit (ref: https://github.com/cogeotiff/rio-tiler-pds/pull/34)
 
 ## 0.4.1 (2020-11-24)